set user scope in User#jwt... but should this be deleted to send it in headers? yes.

+++ Makefile
@@ -1,3 +1,4 @@
-serve:
-	rm src/tmp/pids/server.pid
-	docker-compose up
+go:
+	if [ -e src/tmp/pids/server.pid ] ; then rm src/tmp/pids/server.pid && echo 'removed pid'; fi
+	docker-compose up -d
+	docker attach tipjar_app_1
